+++ interpreter/binary/encode.ml
@@ -77,7 +77,7 @@ struct
   let string bs = len (String.length bs); put_string s bs
   let name n = string (Utf8.encode n)
   let list f xs = List.iter f xs
-  let opt f xo = Lib.Option.app f xo
+  let opt f xo = Option.iter f xo
   let vec f xs = len (List.length xs); list f xs
 
   let gap32 () = let p = pos s in word32 0l; byte 0; p
@@ -109,6 +109,7 @@ struct
 
   let ref_type = function
     | FuncRefType -> s7 (-0x10)
+    | ExnRefType -> s7 (-0x17)
     | ExternRefType -> s7 (-0x11)
 
   let value_type = function
@@ -167,20 +168,8 @@ struct
       op 0x04; block_type bt; list instr es1;
       if es2 <> [] then op 0x05;
       list instr es2; end_ ()
-    | TryCatch (bt, es, ct, ca) ->
-      op 0x06; block_type bt; list instr es;
-      let catch (tag, es) =
-        op 0x07; var tag; list instr es
-      in
-      list catch ct;
-      begin match ca with
-        | None -> ()
-        | Some es -> op 0x19; list instr es
-      end;
-      end_ ()
-    | TryDelegate (bt, es, x) ->
-      op 0x06; block_type bt; list instr es;
-      op 0x18; var x
+    | TryTable (bt, cs, es) ->
+      op 0x1f; block_type bt; vec catch cs; list instr es; end_ ()
     | Br x -> op 0x0c; var x
     | BrIf x -> op 0x0d; var x
     | BrTable (xs, x) -> op 0x0e; vec var xs; var x
@@ -190,7 +179,7 @@ struct
     | ReturnCall x -> op 0x12; var x
     | ReturnCallIndirect (x, y) -> op 0x13; var y; var x
     | Throw x -> op 0x08; var x
-    | Rethrow x -> op 0x09; var x
+    | ThrowRef -> op 0x0a
 
     | Drop -> op 0x1a
     | Select None -> op 0x1b
@@ -745,6 +734,13 @@ struct
     | VecReplace (V128 (I64x2 (V128Op.Replace i))) -> vecop 0x1el; byte i
     | VecReplace (V128 (F32x4 (V128Op.Replace i))) -> vecop 0x20l; byte i
     | VecReplace (V128 (F64x2 (V128Op.Replace i))) -> vecop 0x22l; byte i
+
+  and catch c =
+    match c.it with
+    | Catch (x1, x2) -> byte 0x00; var x1; var x2
+    | CatchRef (x1, x2) -> byte 0x01; var x1; var x2
+    | CatchAll x -> byte 0x02; var x
+    | CatchAllRef x -> byte 0x03; var x
 
   let const c =
     list instr c.it; end_ ()